requirements.txt
azure-storage-blob>=12.0.0
azure-storage-file-share>=12.0.0
azure-storage-file-datalake>=12.0.0
azure-storage-queue>=12.0.0
librosa>=0.11.0
mss==10.0.0
numpy==1.26.4
ollama==0.4.7
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.2.1
pyautogui==0.9.54
Requests==2.32.3
scipy==1.15.2
sounddevice==0.5.1
soundfile==0.13.1
torch==2.5.1
twitchAPI==4.4.0
twitchio==2.10.0
vosk==0.3.44
whisper==1.1.10
